Refactor (#289)
* merge model modules

* Combine end router modules

move version 2 endpoint definitions to single module
move version 1 endpoint definitions to single module

* remove enum subpackage

* add back warning suppression

* fix module names

* add source info to log message

* tweak and fix logging to showup on Heroku
