+++ app/__init__.py
@@ -3,5 +3,9 @@ Corona Virus Tracker API
 ~~~~~~~~~~~~~~~~~~~~~~~~
 API for tracking the global coronavirus (COVID-19, SARS-CoV-2) outbreak.
 """
+import logging
+
 # See PEP396.
 __version__ = "2.0.3"
+
+logging.basicConfig(level=logging.INFO)
