+++ app/services/location/jhu.py
@@ -14,6 +14,8 @@ from ...utils import date as date_util
 from ...utils import httputils
 from . import LocationService
 
+LOGGER = logging.getLogger("services.location.jhu")
+
 
 class JhuLocationService(LocationService):
     """
@@ -48,8 +50,6 @@ async def get_category(category):
     :returns: The data for category.
     :rtype: dict
     """
-    logger = logging.getLogger("services.location.jhu")
-
     # Adhere to category naming standard.
     category = category.lower()
 
@@ -57,15 +57,15 @@ async def get_category(category):
     url = BASE_URL + "time_series_covid19_%s_global.csv" % category
 
     # Request the data
-    logger.info("Requesting data...")
+    LOGGER.info("jhu Requesting data...")
     async with httputils.CLIENT_SESSION.get(url) as response:
         text = await response.text()
 
-    logger.info("Data received")
+    LOGGER.info("jhu Data received")
 
     # Parse the CSV.
     data = list(csv.DictReader(text.splitlines()))
-    logger.info("CSV parsed")
+    LOGGER.info("jhu CSV parsed")
 
     # The normalized locations.
     locations = []
@@ -98,7 +98,7 @@ async def get_category(category):
                 "latest": int(latest or 0),
             }
         )
-    logger.info("Data normalized")
+    LOGGER.info("jhu Data normalized")
 
     # Latest total.
     latest = sum(map(lambda location: location["latest"], locations))
