+++ app/utils/populations.py
@@ -28,7 +28,7 @@ def fetch_populations(save=False):
 
     # Fetch the countries.
     try:
-        countries = requests.get(GEONAMES_URL, params={"username": "dperic"}, timeout=1.5).json()["geonames"]
+        countries = requests.get(GEONAMES_URL, params={"username": "dperic"}, timeout=1.25).json()["geonames"]
         # Go through all the countries and perform the mapping.
         for country in countries:
             mappings.update({country["countryCode"]: int(country["population"]) or None})
@@ -40,6 +40,7 @@ def fetch_populations(save=False):
         mappings = app.io.load(GEONAMES_BACKUP_PATH)
         LOGGER.info(f"Using backup data from {GEONAMES_BACKUP_PATH}")
     # Finally, return the mappings.
+    LOGGER.info("Fetched populations")
     return mappings
